+++ src/com/intellij/idea/plugin/hybris/terminal/HybrisProjectTerminalCustomizer.kt
@@ -1,0 +1,32 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.intellij.idea.plugin.hybris.terminal
+
+import com.intellij.idea.plugin.hybris.settings.components.ProjectSettingsComponent
+import com.intellij.openapi.components.PathMacroManager
+import com.intellij.openapi.project.Project
+import org.jetbrains.plugins.terminal.LocalTerminalCustomizer
+
+class HybrisProjectTerminalCustomizer : LocalTerminalCustomizer() {
+
+    override fun getDefaultFolder(project: Project): String? {
+        if (!ProjectSettingsComponent.getInstance(project).isHybrisProject()) return null
+
+        return PathMacroManager.getInstance(project).expandPath("\$PROJECT_DIR$")
+    }
+}